erpgrpcreport/requirements.txt
Flask>=2.2,<3
grpcio>=1.50,<2
protobuf>=4.21,<6
requests>=2.28,<3
